pyproject.toml
[project]
name = "photoaddwatermark"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "piexif>=1.1.3",
    "pillow>=11.3.0",
]
